Extract Array, Math, NotificationCenter, and IncrementalDelay to WordPressShared
Move the Array and Math extensions and NotificationCenter.observeOnce into
WordPressShared, and split IncrementalDelay out of Delay.swift (leaving the GCD
DispatchDelayedAction/DelayStateWrapper helpers in the app). Relocate their
tests into WordPressSharedTests and add the WordPressShared import to the
remaining callers.
