+++ Modules/Tests/WordPressSharedTests/MathTest.swift
@@ -1,5 +1,5 @@
 import XCTest
-@testable import WordPress
+@testable import WordPressShared
 
 class MathTest: XCTestCase {
