+++ Modules/Tests/WordPressSharedTests/NotificationCenterObserveOnceTests.swift
@@ -1,5 +1,5 @@
 import XCTest
-@testable import WordPress
+@testable import WordPressShared
 
 private let counterKey = "counter"
