+++ WordPress/Classes/Utility/Delay.swift
@@ -1,4 +1,5 @@
 import Foundation
+import WordPressShared
 
 /// A delayed action implemented using GCD
 ///
@@ -58,32 +59,6 @@ struct DispatchDelayedAction {
         init(_ canceled: Bool) {
             self.canceled = canceled
         }
-    }
-}
-
-/// Provides a sequence of incremental delays, repeating the last one
-/// indefinitely.
-///
-public struct IncrementalDelay<Element> {
-    public var current: Element
-
-    private let delaySequence: AnySequence<Element>
-    private var iterator: AnyIterator<Element>
-
-    public init(_ sequence: [Element]) {
-        precondition(!sequence.isEmpty, "IncrementalDelay sequence can't be empty")
-        delaySequence = sequence.repeatingLast()
-        iterator = delaySequence.makeIterator()
-        current = iterator.next()!
-    }
-
-    public mutating func increment() {
-        current = iterator.next()!
-    }
-
-    public mutating func reset() {
-        iterator = delaySequence.makeIterator()
-        current = iterator.next()!
     }
 }
