+++ WordPress/Classes/ViewRelated/Blog/Blog Dashboard/Cards/Activity Log/DashboardActivityLogViewModel.swift
@@ -1,4 +1,5 @@
 import Foundation
+import WordPressShared
 import WordPressKit
 
 class DashboardActivityLogViewModel {
