+++ WordPress/Classes/ViewRelated/Pages/Views/PageListCell.swift
@@ -1,4 +1,5 @@
 import Foundation
+import WordPressShared
 import UIKit
 import Combine
 import AsyncImageKit
