+++ WordPress/Classes/ViewRelated/Post/Search/PostSearchService.swift
@@ -1,4 +1,5 @@
 import CoreData
+import WordPressShared
 import Foundation
 import WordPressData
