+++ WordPress/Classes/ViewRelated/Post/Views/PostListCell.swift
@@ -1,4 +1,5 @@
 import Foundation
+import WordPressShared
 import UIKit
 import AsyncImageKit
 import WordPressData
